ProjectWebsite/mysite/requirements.txt
absl-py==0.13.0
asgiref==3.3.4
astunparse==1.6.3
beepy==1.0.7
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
Django==3.2.4
django-crispy-forms==1.12.0
django-registration-redux==2.9
easydict==1.9
gast==0.3.3
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.38.1
h5py==2.10.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==4.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.2
moviepy==1.0.3
numpy==1.18.5
oauthlib==3.1.1
opencv-python==4.1.1.26
opt-einsum==3.3.0
Pillow==8.2.0
proglog==0.1.9
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
simpleaudio==1.0.4
six==1.16.0
sqlparse==0.4.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0rc0
termcolor==1.1.0
tf-estimator-nightly==2.3.0.dev2020062301
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
